Bases updates on float seconds
- adds updateTime function
- adds logFile in game object for logging
- removes extra thegame-arrow->Run function call

	modified:   Message.cpp
	modified:   MessageQueue.cpp
	modified:   MessageQueue.h
	modified:   actions.cpp
	modified:   base.h
	modified:   game.h
	modified:   hexagon.cpp
	modified:   hexagon.h
	modified:   logic.cpp
	modified:   loop.cpp
	modified:   render.cpp
	modified:   rotator.cpp
	modified:   rotator.h
	modified:   winmain.cpp

+++ MessageQueue.h
@@ -11,7 +11,7 @@ public:
 	~MessageQueue();
 
 	void Render();
-	void Run(DWORD deltaTime);
+	void Run(float deltaTime);
 
 	void AddMessage(string msg, float seconds);
